fix: resolve database FK constraint and cache invalidation issues
- Fix foreign key constraint error by changing BigInteger to Integer in
task_kb_binding.py and alembic migration for column type compatibility

- Fix SQL IN clause parameter binding with bindparam(expanding=True) in queries.py

- Fix RestrictedObserver permission logic in contexts.py to restrict access
if ANY KB is restricted (not just ALL)

- Fix linked_group handling in task_manager.py to only apply to group chats

- Fix _is_kb_bound_to_user_group_chat() with TaskResource join and active check

- Fix useUnifiedMessages.ts cache invalidation by comparing message signatures
instead of just array length

- Fix knowledge_service.py all_kbs query to only apply namespace constraint
to owned KBs, not shared ones

- Fix retrieval_config update to not overwrite retriever_namespace with default

- Extract batch aggregation helpers to knowledge_repository.py module

- Fix MessagesArea.tsx memo comparator to include message.result comparison
for streamed result block updates

+++ backend/alembic/versions/a8b9c0d1e2f3_add_linked_group_id_to_tasks.py
@@ -151,9 +151,9 @@ def _upgrade_task_kb_bindings() -> None:
     # 1. Create the table
     op.create_table(
         "task_knowledge_base_bindings",
-        sa.Column("id", sa.BigInteger(), primary_key=True, autoincrement=True),
-        sa.Column("task_id", sa.BigInteger(), nullable=False),
-        sa.Column("knowledge_base_id", sa.BigInteger(), nullable=False),
+        sa.Column("id", sa.Integer(), primary_key=True, autoincrement=True),
+        sa.Column("task_id", sa.Integer(), nullable=False),
+        sa.Column("knowledge_base_id", sa.Integer(), nullable=False),
         sa.Column("bound_by", sa.String(255), nullable=False),
         sa.Column("bound_at", sa.DateTime(), nullable=False),
         sa.Column(
